iterations/2-poc-gradio/requirements.txt
gradio
haystack-ai
python-dotenv